transferred project files from personal github
transferred project files from personal github to this public github to showcase and honed in my full stack embedded  system skills and experience.  This project is a full stack weather station project using: 
Raspberry Pi and DHT sensor.  The DHT sensor collects the data and then parse it to the Raspberry Pi for processing and then upload them to a basic html page.  The html page was hosted locally on the machine, but it can be posted online if the html was hosted by a web host.